Add streaming support
These changes add streaming support to the crate, so entire files no
longer have to be copied to main memory before being compressed. This
streaming support is safer and easier to use than memmapping the file, a
operation that is not entirely OS-agnostic and requires detailed
analysis of how other processes may interact with the memmapped file to
guarantee stability and safety. The crate now accepts any Read trait
implementation for input data, which provides lots of flexibility.

Compression should be unaffected, as the input data is still read in
blocks of ZOPFLI_MASTER_BLOCK_SIZE like before, and a sliding window of
the last ZOPFLI_WINDOW_SIZE bytes read is kept to initialize the
dictionary for each master block. To test that the changes work okay, I
ran the test/run.sh script, which compressed the test files with no
issues. I got the original idea for this improvement from
https://github.com/google/zopfli/issues/14#issuecomment-77830531.

Although I tried to follow the style of the code that already exists and
not introduce too much modern Rust features, MSRV is now bumped to 1.41.
Also, the public interface exposed by this crate changed in a
backwards-incompatible manner, but the crate major version is still 0,
and semantic versioning conventions tell that public APIs in major
version 0 are unstable, so I don't think this should be a problem.

Resolves #18.