Prevent zopfli from crashing on a single zero-byte file
Correct some logic that was not taking into account the fact that the
final block could also be the first block (with a block length of 0).

This technically closes #2 and prevents zopfli from crashing when given
a single zero-length file as an input or as one of the inputs, except
that I've manually verified a crash still occurs if explicitly using
`BlockType::Uncompressed` or `BlockType::Fixed`. The patch works fine if
using the default `BlockType::Dynamic`. I think those are separate bugs
and would have no bearing on the correctness of this patch, but I'm not
certain since this is my first time looking at the code.