Drop `Seek` or input size knowledge requirements
Requiring input byte sources to implement the `Seek` trait is onerous
for end-users, as most programs that generate DEFLATE streams do not
impose seekability requirements. In the Unix world, it's common to
pipe the output of a program as an input for a compression program,
which is a non-seekable data source. In addition, achieving input
seekability is often non-practical in network scenarios due to buffering
resource requirements and other reasons.

By being smarter about how we use a sliding window, we can drop
seekability and exact input size knowledge requirements from the API
exposed by this crate, making it readily applicable for even more
usage scenarios, without affecting compression. A high-level overview
of the new technique is given through code comments.

A downside I can see of this change is that it requires a sliding window
ZOPFLI_MASTER_BLOCK_SIZE bytes (≈ 1 MiB) bigger than before due to the
need to temporarily store an additional uncompressed master block in
memory. However, I think that the better API design makes this trade-off
worth it: the additional memory usage is insignificant for the kind
of computers that are most likely to run a compression algorithm as
demanding as Zopfli anyway.

This is also a breaking change, as the `compress_seekable` function was
removed from the public API. As a minor note, the sliding window for
the `deflate` function is no longer allocated in the stack, which avoids
running out of stack memory in practical scenarios I've encountered:
allocating ZOPFLI_MASTER_BLOCK_SIZE bytes on the stack is a lot, and the
cost of calling the memory allocator pales in comparison to actually
compressing data with Zopfli.