Expose Zopfli options as public API. Remove verbosity options
As described in #5, users may have good reasons to tweak Zopfli
compression options, but this is not possible with the current API.

To improve the situation, let's refactor the Options struct with proper
data types that guarantee that invariants are being held, and make those
fields public. While at it, let's drop the verbosity options in favor of
using the log crate, which is lightweight and much more flexible and
idiomatic for users.