Verify MSRV in CI workflow
Defining a MSRV policy in a way that does not upset anyone with
reasonable requirements is a hard problem. However, it's even worse to
document a MSRV that does not apply in practice. Let's check that Zopfli
builds fine when using the documented MSRV in CI workflows.

Checking that the beta toolchain builds properly does not provide, in
my opinion, much benefit. Code that builds fine on an old stable Rust
version should build fine on the latest nightly too.