feat: add `GzipEncoder` and `ZlibEncoder`, make `compress` use encoders
This provides a much more flexible and powerful API even in no-std
environments, stepping up the gzip and Zlib formats as we previously did
with raw DEFLATE via `DeflateEncoder`. The new encoder writers are just
wrappers around `DeflateEncoder` that write header and trailer data
before and after the DEFLATE stream, respectively.

While at it, let's remove the `deflate` function in favor of composite
factory methods that return a `DeflateEncoder` wrapped in a `BufWriter`.