Revert commit 94af3b5
Its premise was wrong: the original C code compares the value of
`amount` against `(unsigned short)(-1)`, which is an idiom for an all
ones `unsigned short` value, i.e. `u16::MAX` in Rust, due to the usual
representation of signed integers in two's complement and the cast just
taking the lower bits that fit in a `short`.

This was responsible for breaking Zopfli's hash optimizations for files
with long runs of the same byte, resulting in a 27x slowdown for our
synthetic `master-block-size-zeros.bin` test file. More realistic files
are also likely to be affected, although less dramatically.