Get rid of `ZopfliHash` object pools
Initializing and accessing it has a high cost due to the thread-safe
locking involved, which is especially evident when profiling and
benchmarking the `empty.txt` file against the original Zopfli
implementation, and makes for a somewhat strange prewarming API that
probably nobody knows how to use properly.

The original Zopfli implementation just passes around and reuses an
already allocated hash object instead, which is what we should also be
doing here, unless we want to play unsafe games with statics.