Optimize match prefix length calculation function with autovectorization
On the `calgary-books.txt` file, our crate spent ~2% of its time on the
`get_match` function, easily being among the 10 hottest functions. After
comparing the Rust and C Zopfli source code, it became evident that one
of the reasons for this was the fact that bytes were counted one at a
time for computing match lengths, instead of several at a time, which
significantly reduces looping overhead in modern CPU architectures.

To improve on this, let's use a carefully crafted implementation of that
function that exploits LLVM's ability to portably and cleanly
autovectorize `u128` integers, allowing to analyze up to 2x more bytes
per iteration than the original Zopfli implementation. On the
aforementioned file, this speeds up `get_match` by ~2x, with a fairly
small effect in total compression wall-clock time.