chore: bump MSRV to 1.82
The latest release of `proptest` requires Rust 1.82 at least, so
developers need at least Rust 1.82 to have a proper experience.

Because `proptest` isn't used for downstream consumers of this crate
however, the previous MSRV of 1.73 is still technically appropriate for
most users, as we can define "supported" by "minimum version required
for `cargo build --all-features` to work, which it does.